+++ rotkehlchen/db/upgrade_manager.py
@@ -26,6 +26,7 @@ from rotkehlchen.db.upgrades.v38_v39 import upgrade_v38_to_v39
 from rotkehlchen.db.upgrades.v39_v40 import upgrade_v39_to_v40
 from rotkehlchen.db.upgrades.v40_v41 import upgrade_v40_to_v41
 from rotkehlchen.db.upgrades.v41_v42 import upgrade_v41_to_v42
+from rotkehlchen.db.upgrades.v42_v43 import upgrade_v42_to_v43
 from rotkehlchen.errors.misc import DBUpgradeError
 from rotkehlchen.logging import RotkehlchenLogsAdapter
 from rotkehlchen.utils.interfaces import ProgressUpdater
@@ -105,6 +106,10 @@ UPGRADES_LIST = [
     UpgradeRecord(
         from_version=41,
         function=upgrade_v41_to_v42,
+    ),
+    UpgradeRecord(
+        from_version=42,
+        function=upgrade_v42_to_v43,
     ),
 ]